+++ internal/cmd/cmd_allow.go
@@ -8,10 +8,11 @@ import (
 
 // CmdAllow is `direnv allow [PATH_TO_RC]`
 var CmdAllow = &Cmd{
-	Name:   "allow",
-	Desc:   "Grants direnv to load the given .envrc or .env",
-	Args:   []string{"[PATH_TO_RC]"},
-	Action: actionWithConfig(cmdAllowAction),
+	Name:    "allow",
+	Desc:    "Grants direnv permission to load the given .envrc or .env file.",
+	Args:    []string{"[PATH_TO_RC]"},
+	Aliases: []string{"permit", "grant"},
+	Action:  actionWithConfig(cmdAllowAction),
 }
 
 var migrationMessage = `
@@ -64,7 +65,10 @@ func cmdAllowAction(env Env, args []string, config *Config) (err error) {
 	if err != nil {
 		return err
 	} else if rc == nil {
-		return fmt.Errorf(".envrc or .env file not found")
+		if config.LoadDotenv {
+			return fmt.Errorf(".envrc or .env file not found")
+		}
+		return fmt.Errorf(".envrc file not found")
 	}
 	return rc.Allow()
 }