+++ test/direnv-test-common.sh
@@ -235,13 +235,14 @@ test_start "load-envrc-before-env"
 test_stop
 
 test_start "load-env"
+  echo "[global]
+load_dotenv = true" > "${XDG_CONFIG_HOME}/direnv/direnv.toml"
+  direnv allow
   direnv_eval
   test_eq "${HELLO}" "world"
 test_stop
 
 test_start "skip-env"
-  echo "[global]
-skip_dotenv = true" > "${XDG_CONFIG_HOME}/direnv/direnv.toml"
   direnv_eval
   test -z "${SKIPPED}"
 test_stop
@@ -269,6 +270,15 @@ if has python; then
     fi
   test_stop
 fi
+
+test_start "aliases"
+  direnv deny
+  # check that allow/deny aliases work
+  direnv permit && direnv_eval && test -n "${HELLO}"
+  direnv block  && direnv_eval && test -z "${HELLO}"
+  direnv grant  && direnv_eval && test -n "${HELLO}"
+  direnv revoke && direnv_eval && test -z "${HELLO}"
+test_stop
 
 # Context: foo/bar is a symlink to ../baz. foo/ contains and .envrc file
 # BUG: foo/bar is resolved in the .envrc execution context and so can't find